Question17

+++ ques17.c
@@ -2,7 +2,7 @@
 #include<math.h>
 
 void roots(int a, int b, int c){
-    int d = pow((float)b,2) -(float)(4*a*c);
+    float d = pow((float)b,2) -(float)(4*a*c);
     float r1 = (float)((-b) + sqrt(d))/(2*a);
     float r2 = (float)( (-b) - sqrt(d))/(2*a);
     printf("%f   %f",r1,r2);
